py/requirements.txt
autopep8==1.7.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
pycodestyle==2.9.1
Pygments==2.13.0
python-dotenv==0.21.0
rich==12.5.1
shellingham==1.5.0
toml==0.10.2
typer==0.6.1
